requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
slugify==0.0.1
sqlparse==0.5.3
django-allauth
